requirements.txt
rembg==2.0.67
Pillow==10.0.1
